Increase render timeout due to actual high render times in LaTeX. Refactoring latex element rendering.
