+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/EksempelBrev.kt
@@ -86,13 +86,13 @@ object EksempelBrev : StaticTemplate {
                 text(Bokmal to "Dette er din inntekt fra 01.01.2020 til 01.05.2020")
                 table(
                     header = {
-                        column(3) {text(Bokmal to "Kolonne 1", FontType.BOLD)}
-                        column(1, RIGHT) {text(Bokmal to "Kolonne 2", FontType.BOLD)}
-                        column(1, RIGHT) {text(Bokmal to "Kolonne 3", FontType.BOLD)}
-                        column(1, RIGHT) {text(Bokmal to "Kolonne 4", FontType.BOLD)}
+                        column(3) { text(Bokmal to "Kolonne 1", FontType.BOLD) }
+                        column(1, RIGHT) { text(Bokmal to "Kolonne 2", FontType.BOLD) }
+                        column(1, RIGHT) { text(Bokmal to "Kolonne 3", FontType.BOLD) }
+                        column(1, RIGHT) { text(Bokmal to "Kolonne 4", FontType.BOLD) }
                     }
                 ) {
-                    for (i in 1..50) {
+                    for (i in 1..10) {
                         row {
                             cell {
                                 text(Bokmal to "Dette er ditt beløp av noe test test test test test test test test")
