+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/template/base/PensjonLatex.kt
@@ -6,9 +6,6 @@ import no.nav.pensjon.brev.api.model.SignerendeSaksbehandlere
 import no.nav.pensjon.brev.latex.LatexPrintWriter
 import no.nav.pensjon.brev.model.format
 import no.nav.pensjon.brev.template.*
-import no.nav.pensjon.brev.template.Element.Table.ColumnAlignment.LEFT
-import no.nav.pensjon.brev.template.Element.Table.ColumnAlignment.RIGHT
-import no.nav.pensjon.brev.template.Element.Text.FontType.*
 import no.nav.pensjon.brev.template.base.pensjonlatex.pensjonLatexSettings
 import java.time.LocalDate
 import java.time.ZonedDateTime
@@ -188,150 +185,4 @@ object PensjonLatex : BaseTemplate() {
         letter.template.outline.forEach { renderElement(scope, it, printWriter) }
     }
 
-    private fun renderElement(scope: ExpressionScope<*, *>, element: Element<*>, printWriter: LatexPrintWriter): Unit =
-        when (element) {
-            is Element.Title1 ->
-                with(printWriter) {
-                    printCmd("lettersectiontitle") {
-                        arg { element.title1.forEach { child -> renderElement(scope, child, it) } }
-                    }
-                }
-
-            is Element.IncludePhrase<*, *> -> {
-                val newScope = ExpressionScope(element.data.eval(scope), scope.felles, scope.language)
-                element.phrase.elements.forEach { renderElement(newScope, it, printWriter) }
-            }
-
-            is Element.Conditional ->
-                with(element) {
-                    val toRender = if (predicate.eval(scope)) showIf else showElse
-                    toRender.forEach { renderElement(scope, it, printWriter) }
-                }
-
-            is Element.Text.Literal -> {
-                val textLiteral = element.text(scope.language)
-                with(printWriter) {
-                    when (element.fontType) {
-                        PLAIN -> print(textLiteral)
-                        BOLD -> printCmd("textbf") { arg { print(textLiteral) } }
-                        ITALIC -> printCmd("textit") { arg { print(textLiteral) } }
-                    }
-                }
-
-            }
-
-            is Element.Text.Expression ->
-                printWriter.print(element.expression.eval(scope))
-
-            is Element.Text.Expression.ByLanguage ->
-                printWriter.print(element.expr(scope.language).eval(scope))
-
-            is Element.ItemList ->
-                with(printWriter) {
-                    val items = element.items.filter { it.condition == null || it.condition.eval(scope) }
-                    if (items.isEmpty()) return
-                    printCmd("begin") {
-                        arg { print("itemize") }
-                    }
-                    items.forEach { item ->
-                        print("""\item """, escape = false)
-                        item.elements.forEach {
-                            renderElement(scope, it, this)
-                        }
-                    }
-                    printCmd("end") {
-                        arg { print("itemize") }
-                    }
-                }
-            is Element.Paragraph ->
-                printWriter.printCmd("templateparagraph") {
-                    arg { element.paragraph.forEach { child -> renderElement(scope, child, it) } }
-                }
-
-            is Element.Form.Text ->
-                with(printWriter) {
-                    if (element.vspace) {
-                        printCmd("formvspace")
-                    }
-
-                    printCmd("formText") {
-                        arg {
-                            renderElement(scope, element.prompt, it)
-                            it.print(" ${".".repeat(element.size)}")
-                        }
-                    }
-                }
-
-            is Element.Form.MultipleChoice ->
-                with(printWriter) {
-                    if (element.vspace) {
-                        printCmd("formvspace")
-                    }
-
-                    printCmd("begin") {
-                        arg { it.print("formChoice") }
-                        arg { renderElement(scope, element.prompt, it) }
-                    }
-
-                    element.choices.forEach {
-                        printCmd("item")
-                        renderElement(scope, it, printWriter)
-                    }
-
-                    printCmd("end", "formChoice")
-                }
-
-            is Element.NewLine ->
-                printWriter.printCmd("newline")
-            is Element.Table ->
-                with(printWriter) {
-                    val rows = element.rows.filter { it.condition == null || it.condition.eval(scope) }
-                    if (rows.isEmpty()) return
-                    val columnSpec = element.header.colSpec
-                    printCmd("begin") {
-                        arg { print("letterTable") }
-                        arg { print(columnHeadersLatexString(columnSpec)) }
-                    }
-
-                    renderTableCells(columnSpec.map { it.headerContent }, scope, printWriter, columnSpec)
-                    rows.forEach { renderTableCells(it.cells, scope, printWriter, columnSpec) }
-
-                    printCmd("end") { arg { print("letterTable") } }
-                }
-        }
-
-    private fun columnHeadersLatexString(columnSpec: List<Element.Table.ColumnSpec<out LanguageSupport>>) =
-        columnSpec.map {
-            ("X" +
-                    when (it.alignment) {
-                        LEFT -> "[l]"
-                        RIGHT -> "[r]"
-                    }).repeat(it.columnSpan)
-        }.joinToString("")
-
-    private fun renderTableCells(
-        cells: List<Element.Table.Cell<out LanguageSupport>>,
-        scope: ExpressionScope<*, *>,
-        printWriter: LatexPrintWriter,
-        colSpec: List<Element.Table.ColumnSpec<out LanguageSupport>>
-    ) {
-        with(printWriter) {
-            cells.forEachIndexed { index, cell ->
-                val columnSpan = colSpec[index].columnSpan
-                if (columnSpan > 1) {
-                    print("\\SetCell[c=$columnSpan]{}", escape = false)
-                }
-                cell.elements.forEach { cellElement ->
-                    renderElement(scope, cellElement, printWriter)
-                }
-                if (columnSpan > 1) {
-                    print(" ${"& ".repeat(columnSpan - 1)}", escape = false)
-                }
-                if (index < cells.lastIndex) {
-                    print("&", escape = false)
-                }
-            }
-            print("""\\""", escape = false)
-        }
-    }
 }
